warnings fixes

+++ src/RSArea.cpp
@@ -287,13 +287,13 @@ void RSArea::ParseTriFile(PakEntry *entry) {
             v->u2 = stream.ReadByte();
             coo = stream.ReadInt24LE();
             read += 4;
-            v->x = (float) coo * BLOCK_COORD_SCALE;
+            v->x = coo * (int) BLOCK_COORD_SCALE;
             coo = stream.ReadInt24LE();
             read += 4;
-            v->z = (float) coo * BLOCK_COORD_SCALE;
+            v->z = coo * (int) BLOCK_COORD_SCALE;
             coo = stream.ReadShort();
             read += 2;
-            v->y = (float) coo * HEIGH_MAP_SCALE;
+            v->y = coo * (int) HEIGH_MAP_SCALE;
             overTheMapIsTheRunway.lx = ((overTheMapIsTheRunway.lx == 0) && (i == 0)) ? v->x : overTheMapIsTheRunway.lx;
             overTheMapIsTheRunway.hx = ((overTheMapIsTheRunway.hx == 0) && (i == 0)) ? v->x : overTheMapIsTheRunway.hx;
             overTheMapIsTheRunway.ly = ((overTheMapIsTheRunway.ly == 0) && (i == 0)) ? v->z : overTheMapIsTheRunway.ly;
